server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["The <at.thealchemist@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.5"
django-configurations = "^2.2"
django-constance = "^2.8.0"
psycopg2-binary = "^2.9.1"
gunicorn = "^20.1.0"
uvicorn = "^0.14.0"
celery = "^5.1.2"
redis = "^3.5.3"
django-celery-beat = "^2.2.1"
vk = "^2.0.2"
vk-api = "^11.9.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
